feat: add types to order-feed page
